requirements.txt
streamlit==1.32.2
pandas==2.2.1
altair==5.2.0
fpdf2==2.7.8
